backend/requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx==0.2.4
idna==3.7
lxml==5.2.2
paypalrestsdk==1.13.3
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.0
stripe==10.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
